Implement more flexible filtering for the store
Ember Data's filtering has rather limited capabilities. The biggest
problem is that it will check if a record should be added to a filterred
array only when a record is added to the store. This sometimes creates
problems like with starred repositories. We want to update the starred
repositories list when a repository is starred and that's why I had to
implement starredRepos service. The new filtering mechanism allows to
make it simple with one change to the store's API - fourth argument
being an array of dependencies which trigger a refresh of the filtered
array.